Checkpoint, processed 10 minute video into ~60s chunks.
