+++ lib/media/video.dart
@@ -1,6 +1,5 @@
 import 'dart:convert';
-import 'dart:io';
-import 'dart:math';
+
 import 'package:flutter/material.dart' as mat;
 import 'package:flutter_fhe_video_similarity/logging.dart';
 import 'package:flutter_fhe_video_similarity/media/cache.dart';
@@ -9,239 +8,10 @@ import 'package:opencv_dart/opencv_dart.dart' as cv;
 import 'package:flutter/foundation.dart';
 
 import 'processor.dart';
-import 'primatives.dart';
 import 'image.dart';
 
-class VideoMeta extends Meta {
-  String codec;
-  int fps;
-  Duration duration;
-  String sha256;
-  int startFrame;
-  int endFrame;
-  int totalFrames;
-  String encryptionStatus;
-  late String id; // Unique identifier of Video
-
-  VideoMeta(
-      {required this.codec,
-      required this.fps,
-      required this.totalFrames,
-      required this.duration,
-      required this.sha256,
-      required this.startFrame,
-      required this.endFrame,
-      required this.encryptionStatus,
-      required String name,
-      required String extension,
-      required DateTime created,
-      required DateTime modified,
-      required String path})
-      : super(name, extension, created, modified, path) {
-    id = '${sha256.substring(0, 8)}-${created.millisecondsSinceEpoch}';
-  }
-
-  @override
-  Map toJson() => {
-        ...super.toJson(),
-        'codec': codec,
-        'fps': fps,
-        'totalFrames': totalFrames,
-        'sha256': sha256,
-        'durationSeconds': duration.inSeconds,
-        'startFrame': startFrame,
-        'endFrame': endFrame,
-        'encryptionStatus': encryptionStatus,
-      };
-
-  VideoMeta.fromJson(Map<String, dynamic> json)
-      : this(
-            codec: json['codec'],
-            fps: json['fps'],
-            totalFrames: json['totalFrames'],
-            duration: Duration(seconds: json['durationSeconds']),
-            sha256: json['sha256'],
-            startFrame: json['startFrame'],
-            endFrame: json['endFrame'],
-            name: json['name'],
-            extension: json['extension'],
-            created: DateTime.parse(json['created']),
-            modified: DateTime.parse(json['modified']),
-            encryptionStatus: json['encryptionStatus'],
-            path: json['path']);
-
-  VideoMeta.fromFile(File file)
-      : this.fromJson(jsonDecode(file.readAsStringSync()));
-
-  void pprint() {
-    print('--- Video Information ---');
-    print(' * Codec: $codec');
-    print(' * Duration: $duration seconds');
-    print(' * FPS: $fps');
-    print(' * Frame Count: $totalFrames');
-    print(' * Start Frame: $startFrame');
-    print(' * End Frame: $endFrame');
-    print(' * Path: $path');
-    print('-------------------------');
-  }
-
-  @override
-  String toString() => toJson().toString();
-}
-
-/// Get the OpenCV API preference based on the platform
-///
-int get cvApiPreference {
-  if (Platform.isAndroid) {
-    return cv.CAP_ANDROID;
-  } else {
-    return cv.CAP_FFMPEG;
-  }
-}
-
-enum ImageFormat { jpg, png }
-
-enum FrameCount { all, even, odd, firstLast, random5, randomHalf }
-
-class FrameData {
-  final int id;
-  final Uint8List bytes;
-
-  FrameData(this.id, this.bytes);
-}
-
-class FrameIsolate {
-  final List<int> frameIds;
-  final String videoPath;
-  String frameFormat = 'png';
-
-  FrameIsolate(this.frameIds, this.videoPath);
-}
-
-Future<List<Uint8List>> extractFrames(FrameIsolate args) async {
-  final bytes = <Uint8List>[];
-
-  // Open the video file in the isolate
-  final copy = await cv.VideoCaptureAsync.fromFileAsync(args.videoPath,
-      apiPreference: cvApiPreference);
-  var success = await copy.grabAsync();
-  if (!copy.isOpened || args.frameIds.isEmpty || !success) return [];
-
-  var idx = 0;
-  // Iterate through 0..N inclusively, assumes frameIds are sorted
-  while (success && idx <= args.frameIds.last) {
-    if (args.frameIds.contains(idx)) {
-      var (exist, image) = await copy.readAsync();
-      if (!image.isEmpty) {
-        bytes.add((await cv.imencodeAsync(".${args.frameFormat}", image)).$2);
-      }
-    }
-    success = await copy.grabAsync();
-    idx += 1;
-  }
-  await copy.releaseAsync();
-
-  return bytes;
-}
-
-/// Extract the frame sizes of the video
-///
-Future<List<int>> extractFrameSizes(FrameIsolate args) async {
-  final sizes = <int>[];
-
-  // Open the video file in the isolate
-  final copy = await cv.VideoCaptureAsync.fromFileAsync(args.videoPath,
-      apiPreference: cvApiPreference);
-  var success = await copy.grabAsync();
-  if (!copy.isOpened || args.frameIds.isEmpty || !success) return [];
-
-  var idx = 0;
-  // Iterate through 0..N inclusively, assumes frameIds are sorted
-  while (success && idx <= args.frameIds.last) {
-    if (args.frameIds.contains(idx)) {
-      var (exist, image) = await copy.readAsync();
-      if (!image.isEmpty) {
-        var frameData =
-            (await cv.imencodeAsync(".${args.frameFormat}", image)).$2;
-        sizes.add(frameData.length);
-      }
-    }
-    success = await copy.grabAsync();
-    idx += 1;
-  }
-  await copy.releaseAsync();
-  return sizes;
-}
-
-/// Step through the video frame by frame and count the number of frames
-///
-/// Exposing this method to the user is not recommended as it is slow and
-/// inefficient. Use [frameCount] instead.
-///
-int countAllFrames(String videoPath) {
-  // Create + Release video copy
-  final copy = cv.VideoCapture.fromFile(videoPath);
-  var frames = 0;
-  var (success, _) = copy.read();
-  while (success) {
-    frames += 1;
-    (success, _) = copy.read();
-  }
-  copy.release();
-  return frames;
-}
-
-/// Estimate the timestamps at the start of each segment
-///
-List<DateTime> timestampsFromSegment(VideoMeta meta, Duration segmentDuration) {
-  List<DateTime> segments = [];
-  for (int i = meta.startFrame;
-      i < meta.endFrame;
-      i += segmentDuration.inSeconds * meta.fps) {
-    segments.add(meta.created.add(Duration(seconds: i ~/ meta.fps)));
-  }
-  return segments;
-}
-
-/// Generate ranges of frame indices based on the [segmentDuration]
-///
-List<List<int>> frameIndexFromSegment(
-    VideoMeta meta, Duration segmentDuration, FrameCount frameCount) {
-  List<List<int>> segments = [];
-  int segment = segmentDuration.inSeconds * meta.fps;
-
-  for (int i = meta.startFrame; i < meta.endFrame; i += segment) {
-    // Drop all values greater than the end frame
-    var inc = List<int>.generate(segment, (inc) => i + inc);
-    var all = inc.where((idx) => idx <= meta.endFrame).toList();
-
-    switch (frameCount) {
-      case FrameCount.firstLast:
-        segments.add([all.first, all.last]);
-        break;
-      case FrameCount.even:
-        segments.add(all.where((idx) => idx % 2 == 0).toList());
-        break;
-      case FrameCount.odd:
-        segments.add(all.where((idx) => idx % 2 != 0).toList());
-        break;
-      case FrameCount.all:
-        segments.add(all);
-        break;
-      case FrameCount.random5:
-        final allRandom = all;
-        allRandom.shuffle();
-        segments.add(allRandom.take(5).toList());
-        break;
-      case FrameCount.randomHalf:
-        final halfRandom = all;
-        halfRandom.shuffle(Random());
-        segments.add(halfRandom.take(all.length ~/ 2).toList());
-        break;
-    }
-  }
-  return segments;
-}
+export 'package:flutter_fhe_video_similarity/media/video_opencv.dart';
+export 'package:flutter_fhe_video_similarity/media/video_meta.dart';
 
 class Video extends UploadedMedia {
   // Logger log = Logger('Video');
@@ -259,7 +29,7 @@ class Video extends UploadedMedia {
     created = timestamp;
 
     // Get frame count
-    totalFrames = frameCount;
+    totalFrames = frameCount(file.path);
 
     // Calculate the number of 'extra' frames exceeding the absolute second
     var extraFrames = (totalFrames % fps).toInt();
@@ -302,10 +72,11 @@ class Video extends UploadedMedia {
       '$hash/$startFrame-$endFrame-${created.millisecondsSinceEpoch}';
 
   Future<void> cache() async {
-    final bytes = await asBytes; // TODO: trim?
+    final bytes = await asBytes;
     manifest.write(bytes.toList(), pwd, "video.mp4");
 
-    stats.cache(pwd);
+    final List<int> metaBytes = utf8.encode(jsonEncode(stats.toJson())).toList();
+    manifest.write(metaBytes, pwd, "meta.json");
   }
 
   int get fps => video.get(cv.CAP_PROP_FPS).toInt();
@@ -347,39 +118,13 @@ class Video extends UploadedMedia {
         "Final frame range after trimming: startFrame=$startFrame, endFrame=$endFrame");
   }
 
-  /// Calculate the number of frames in the video
-  ///
-  /// The number of frames is calculated using the CAP_PROP_FRAME_COUNT property
-  /// of the video. If the property is not supported by the codec, the number of
-  /// frames is calculated manually.
-  ///
-  int get frameCount {
-    final frames = video.get(cv.CAP_PROP_FRAME_COUNT).toInt();
-
-    // CAP_PROP_FRAME_COUNT is not supported by all codecs
-    if (frames == 0) {
-      return countAllFrames(xfile.path);
-    }
-    return frames;
-  }
-
-  Future<List<int>> probeFrameSizes({List<int> frameIds = const [0]}) async {
-    return await compute<FrameIsolate, List<int>>(
-        extractFrameSizes, FrameIsolate(frameIds, xfile.path));
-  }
-
-  Future<List<Uint8List>> probeFrames({List<int> frameIds = const [0]}) async {
-    return await compute<FrameIsolate, List<Uint8List>>(
-        extractFrames, FrameIsolate(frameIds, xfile.path));
-  }
-
   /// Generate an [Image] from the video
   ///
   /// A thumbnail is generated from the frame at [frameIdx] of the video.
   ///
   Future<Image> thumbnail(String filename, [frameIdx = 0]) async {
     Uint8List frameFromIndex =
-        await probeFrames(frameIds: [frameIdx]).then((frames) => frames.first);
+        await probeFrames(FrameIsolate([frameIdx], xfile.path)).then((frames) => frames.first);
     Uint8List resizedFrame = resize(frameFromIndex, ImageFormat.png, 500, 500);
 
     return Image.fromBytes(resizedFrame, created, xfile.path, filename);
